requirements.txt
streamlit
pandas
python-dotenv
azure-ai-formrecognizer
openpyxl
Pillow
